UserNotify dump/restore
